feat: move navbar component and use it on start page
